twkevinzhang_komica/app/src/requirements.txt
protobuf
grpcio===1.67.1
